facebook-sentiment-analyzer/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
facebook-scraper==0.2.59
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
pillow==10.1.0
opencv-python==4.8.1.78
transformers==4.35.2
torch==2.1.1
matplotlib==3.8.2
jupyter==1.0.0
webdriver-manager==4.0.1
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
